+++ kite-web/src/pages/apps/[appId]/events/index.tsx
@@ -1,5 +1,4 @@
 import AppLayout from "@/components/app/AppLayout";
-import AppLayoutV2 from "@/components/app/AppLayoutV2";
 import EventListenerList from "@/components/app/EventListerList";
 import { Separator } from "@/components/ui/separator";
 import env from "@/lib/env/client";
@@ -12,7 +11,7 @@ const breadcrumbs = [
 
 export default function AppEventsPage() {
   return (
-    <AppLayoutV2 title="Event Listeners" breadcrumbs={breadcrumbs}>
+    <AppLayout title="Event Listeners" breadcrumbs={breadcrumbs}>
       <div>
         <h1 className="text-lg font-semibold md:text-2xl mb-1">
           Event Listeners
@@ -30,6 +29,6 @@ export default function AppEventsPage() {
       </div>
       <Separator className="my-8" />
       <EventListenerList />
-    </AppLayoutV2>
+    </AppLayout>
   );
 }
