+++ kite-web/src/pages/apps/[appId]/settings/index.tsx
@@ -1,12 +1,11 @@
 import AppLayout from "@/components/app/AppLayout";
-import { Separator } from "@/components/ui/separator";
 import AppSettingsAppearance from "@/components/app/AppSettingsAppearance";
 import AppSettingsCollaborators from "@/components/app/AppSettingsCollaborators";
+import AppSettingsControls from "@/components/app/AppSettingsControls";
 import AppSettingsCredentials from "@/components/app/AppSettingsCredentials";
 import AppSettingsDelete from "@/components/app/AppSettingsDelete";
 import AppSettingsPresence from "@/components/app/AppSettingsPresence";
-import AppSettingsControls from "@/components/app/AppSettingsControls";
-import AppLayoutV2 from "@/components/app/AppLayoutV2";
+import { Separator } from "@/components/ui/separator";
 
 const breadcrumbs = [
   {
@@ -16,7 +15,7 @@ const breadcrumbs = [
 
 export default function AppSettingsPage() {
   return (
-    <AppLayoutV2 title="App Settings" breadcrumbs={breadcrumbs}>
+    <AppLayout title="App Settings" breadcrumbs={breadcrumbs}>
       <div className="flex flex-col md:flex-row justify-between items-end space-y-5 md:space-y-0">
         <div>
           <h1 className="text-lg font-semibold md:text-2xl mb-1">
@@ -37,6 +36,6 @@ export default function AppSettingsPage() {
         <AppSettingsCollaborators />
         <AppSettingsDelete />
       </div>
-    </AppLayoutV2>
+    </AppLayout>
   );
 }
