+++ kite-web/src/pages/apps/[appId]/variables/index.tsx
@@ -1,5 +1,4 @@
 import AppLayout from "@/components/app/AppLayout";
-import AppLayoutV2 from "@/components/app/AppLayoutV2";
 import VariableList from "@/components/app/VariableList";
 import { Separator } from "@/components/ui/separator";
 import env from "@/lib/env/client";
@@ -11,7 +10,7 @@ const breadcrumbs = [
 
 export default function AppVariablesPage() {
   return (
-    <AppLayoutV2 title="Stored Variables" breadcrumbs={breadcrumbs}>
+    <AppLayout title="Stored Variables" breadcrumbs={breadcrumbs}>
       <div>
         <h1 className="text-lg font-semibold md:text-2xl mb-1">
           Stored Variables
@@ -31,6 +30,6 @@ export default function AppVariablesPage() {
       </div>
       <Separator className="my-8" />
       <VariableList />
-    </AppLayoutV2>
+    </AppLayout>
   );
 }
